==> 00. PII Cleaning Reporting.ipynb <==
## Imports

# Libraries
import pandas as pd
import numpy as np
import os
import hashlib

# Files
path = os.path.expanduser('~/Desktop/CareerFoundry/2.4/4.10.') # path to the root folder

### Orders
df_orders = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv')) 
### Customers
df_customers = pd.read_csv(os.path.join(path, '02 Data', 'Original Data','customers.csv'))

## PII Cleaning and Anonymization

- **First/Last Names** were dropped as direct identifiers.
- **User IDs** were SHA-256 hashed to ensure anonymity while preserving the ability to track behavior across datasets.
- **Age** was converted into broad age groups.
- **Income** was bucketed using quartiles to preserve relative distribution without exposing raw values.
- **Date Joined** was shifted from %m/%d/%Y format to %m/%Y to remove tracability.
These steps follow data privacy best practices while preserving analytical value for segmentation and modeling.

## Drop direct identifiers 

df_customers.info()

df_customers = df_customers.drop(columns=['First Name', 'Surnam'])

## Hash user_id

df_orders.info()

def hash_id(val):
    return hashlib.sha256(str(val).encode()).hexdigest()

df_customers['user_id'] = df_customers['user_id'].apply(hash_id)

df_orders['user_id'] = df_orders['user_id'].apply(hash_id)

## Bucket ages (into 10-year ranges)

df_customers['age_group'] = pd.cut(
    df_customers['Age'],
    bins=[0, 18, 25, 35, 45, 55, 65, np.inf],
    labels=['0-18', '19-25', '26-35', '36-45', '46-55', '56-65', '65+']
)

df_customers = df_customers.drop(columns=['Age'])

## Bucket yearly income into quartiles

df_customers['income_bracket'] = pd.qcut(
    df_customers['income'], 
    q=4, 
    labels=['Low', 'Mid-low', 'Mid-high', 'High']
)

df_customers = df_customers.drop(columns=['income'])

## Change date_joined format to remove exact traceability

df_customers['date_joined'] = pd.to_datetime(df_customers['date_joined'])

df_customers['date_joined'] = df_customers['date_joined'].dt.strftime('%m/%Y') # changing from %m/%d/%Y to %m/%Y

df_orders.to_csv(os.path.join(path, '02 Data','Original Data', 'products.csv'))

df_customers.to_csv(os.path.join(path, '02 Data','Original Data', 'customers.csv'))
