==> 01. Data Cleaning and Consistency Checks.ipynb <==
# Imports

# Libraries
import pandas as pd
import numpy as np
import os

# Files
path = os.path.expanduser('~/Desktop/CareerFoundry/2.4/4.10.') # path to the root folder

### Products
df_products = pd.read_csv(
    os.path.join('~/Desktop/CareerFoundry/Data/products.csv'),
    index_col=False
)
### Orders
df_orders = pd.read_csv(os.path.join(path, '02 Data','Original Data', 'orders.csv')) 
### Customers
df_customers = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))
### Departments
df_departments = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'))
### Orders Prior
df_orders_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'))

## Cleaning products.csv

### Products
df_products.info()

df_products.head()

### 3. Datatype Consistency

df_products = df_products.astype({'product_id':'str', 'aisle_id':'category', 'department_id':'category'})

### 4. Duplicates

print(df_products.duplicated().sum()) 

df_products = df_products.drop_duplicates()

### 5. Nulls

df_products.isnull().sum()

df_products = df_products[df_products['product_name'].isnull() == False]

## Cleaning orders.csv

df_orders.head()

### 1. Droping Unnecessary Columns

### Orders
df_orders = df_orders.drop(columns=['eval_set'])

### 2. Renaming Columns

df_orders.rename(columns = {'order_dow':'order_day_of_week'}, inplace=True)

### 3. Datatype Consistency

df_orders.info()

df_orders = df_orders.astype({'order_id':'str', 'user_id':'str'})

### 4. Duplicates

print(df_orders.duplicated().sum()) # no duplicates found

### 5. Nulls

df_orders.isnull().sum()

df_orders = df_orders.fillna(0) # Nulls probably mean that the customer made the first order, so they are replaced with '0'

## Cleaning customers.csv

df_customers.head()

### 2. Renaming Columns

df_customers.rename(columns = { 'Gender':'gender', 'STATE':'state', 'n_dependants':'dependants',
                               'fam_status':'family_status', 
                    'income':'yearly_income'}, inplace=True)

### 3. Datatype Consistency

df_customers.info()

df_customers=df_customers.astype({'user_id':'str', 'gender':'category', 'state':'category', 'date_joined':'datetime64[ns]', 'family_status':'category'})

### 4. Duplicates

print(df_customers.duplicated().sum()) # no duplicates

### 5. Nulls

df_customers.isnull().sum() # there is no need to remove recors with NULLs as first name is not important for the analysis

## Cleaning departments.csv

df_departments.head() # Data is in the wide format. 

### 0. Transposing data

df_departments = df_departments.T

df_departments.reset_index()

new_header = df_departments.iloc[0]

df_departments = df_departments[1:]

df_departments.columns = new_header

df_departments.head()

### 4. Duplicates

print(df_departments.duplicated().sum()) 

## Cleaning orders_products_prior.csv

df_orders_prior.head()

### 3. Datatype Consistency

df_orders_prior.info()

df_orders_prior = df_orders_prior.astype({'order_id':'str', 'product_id':'str'})

### 4. Duplicates

print(df_orders_prior.duplicated().sum()) # no duplicates found

### 5. Nulls

df_orders_prior.isnull().sum()

# Exports

Products

df_products.to_csv(os.path.join(path, '02 Data','Clean Data', 'products.csv'))

df_products.to_pickle(os.path.join(path, '02 Data','Clean Data', 'products.pkl'))

Orders

df_orders.to_csv(os.path.join(path, '02 Data','Clean Data', 'orders.csv'))

df_orders.to_pickle(os.path.join(path, '02 Data','Clean Data', 'orders.pkl'))

Customers

df_customers.to_csv(os.path.join(path, '02 Data','Clean Data', 'customers.csv'))

df_customers.to_pickle(os.path.join(path, '02 Data','Clean Data', 'customers.pkl'))

Departments

df_departments.to_csv(os.path.join(path, '02 Data','Clean Data', 'departments.csv'))

df_departments.to_pickle(os.path.join(path, '02 Data','Clean Data', 'departments.pkl'))

Prior Orders

df_orders_prior.to_csv(os.path.join(path, '02 Data','Clean Data', 'orders_products_prior.csv'))

df_orders_prior.to_pickle(os.path.join(path, '02 Data','Clean Data', 'orders_products_prior.pkl'))
