==> 03. Combining Exporting.ipynb <==
# 1. Imports

# Libraries
import pandas as pd
import numpy as np
import os
import hashlib

# Files
path = os.path.expanduser('~/Desktop/CareerFoundry/2.4/4.10.') # path to the root folder

### Products
df_products = pd.read_pickle(os.path.join(path, '02 Data', 'Clean Data', 'products.pkl'))
### Orders
df_orders = pd.read_pickle(os.path.join(path, '02 Data', 'Clean Data', 'orders.pkl')) 
### Customers
df_customers = pd.read_pickle(os.path.join(path, '02 Data', 'Clean Data','customers.pkl'))
### Departments
df_departments = pd.read_pickle(os.path.join(path, '02 Data', 'Clean Data', 'departments.pkl'))
### Prior Orders
df_orders_prior= pd.read_pickle(os.path.join(path, '02 Data', 'Clean Data', 'orders_products_prior.pkl'))

## Orders and Products Prior = df_orders_products_prior_merged

df_orders.head()

df_orders_products_prior_merged = []
df_orders_products_prior_merged = df_orders.merge(df_orders_prior, on='order_id', indicator = 'merge_flag_orders_prod_prior')

df_orders_products_prior_merged.head()

df_orders_products_prior_merged['merge_flag_orders_prod_prior'].value_counts() # checking how data was merged

## Products and Orders Products Prior Merged = df_orders_products_prior_products_merged

df_orders_products_prior_products_merged = []
df_orders_products_prior_products_merged = df_orders_products_prior_merged.merge(df_products, on = ['product_id'], 
                                                                indicator='merge_flag_orders_products_prior_products')

df_orders_products_prior_products_merged['merge_flag_orders_products_prior_products'].value_counts() # checking how data was merged

## Customers and Orders Products Prior Merged products = df_order_history

def hash_id(val): # because we hashed user ids in Step 0, we need to prepare data to be able to use it for merging.
    return hashlib.sha256(str(val).encode()).hexdigest()

# decoding hashed values in each df, so that they could be compared
df_customers['user_id_hashed'] = df_customers['user_id'].apply(hash_id)
df_orders_products_prior_products_merged['user_id_hashed'] = df_orders_products_prior_products_merged['user_id'].apply(hash_id)

df_customers.info()

df_orders_products_prior_products_merged.head()

df_order_history = df_orders_products_prior_products_merged.merge(
    df_customers,
    on='user_id',
    how='left',  # Use left join to avoid explosion
    indicator='merge_flag_customer'
)

df_order_history.head()

df_order_history['merge_flag_customer'].value_counts() # checking how data was merged

print(df_order_history.duplicated().sum()) # duplicates check 

## Imports

df_order_history.to_csv(os.path.join(path, '02 Data','Clean Data', 'order_history.csv'))

df_order_history.to_pickle(os.path.join(path, '02 Data','Clean Data', 'order_history.pkl'))

df_orders_products_prior_products_merged.to_csv(os.path.join(path, '02 Data','Clean Data', 'order_history_no_customer.csv'))

df_orders_products_prior_products_merged.to_pickle(os.path.join(path, '02 Data','Clean Data', 'order_history_no_customer.pkl'))
