==> 05. Visualizations.ipynb <==
### 1. Imports

# Libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import hashlib

# Files
path = os.path.expanduser('~/Desktop/CareerFoundry/2.4/4.10.') # path to the root folder

### All Instacart Data Merged
df_customers_orders_products_merged = pd.read_pickle(os.path.join(
    path, '02 Data','Cleaned Data', 'df_customers_orders_products_merged.pkl'))
###

df_customers = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

### 2. Hashing Sensitive Data 

def hash_value(val): # function to hash sensitive values
    return hashlib.sha256(str(val).encode()).hexdigest()

# Hashing df_customers_orders_products_merged
df_customers_orders_products_merged['hashed_user'] = df_customers_orders_products_merged.apply(lambda x: 
                                                    hash_value(x['first_name'] + x['last_name']), axis=1)

# Drop original PII
df_customers = df_customers.drop(columns=['first_name', 'last_name'])

# Hashing df_customers_orders_products_merged
df_customers['hashed_user'] = df_customers.apply(lambda x: hash_value(x['first_name'] + x['last_name']), axis=1)

# Drop original PII
df_customers = df_customers.drop(columns=['first_name', 'last_name'])
